Update display on trade cli
Origin: develop

+++ src/main/java/net/avdw/eve/domain/TradeStatistic.java
@@ -1,13 +1,16 @@
 package net.avdw.eve.domain;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 public class TradeStatistic {
-    public Long volume;
-    public Double minPrice;
-    public Double fivePercentPrice;
-    public Double maxPrice;
-    public Double median;
-    public Double average;
-    public Double weightedAverage;
-    public Double variance;
-    public Double standardDeviation;
+    public BigInteger volume;
+    public BigDecimal minPrice;
+    public BigDecimal fivePercentPrice;
+    public BigDecimal maxPrice;
+    public BigDecimal median;
+    public BigDecimal average;
+    public BigDecimal weightedAverage;
+    public BigDecimal variance;
+    public BigDecimal standardDeviation;
 }
